Make the projects compatible with VS 2008. * Change std::unique_ptr to UniquePtr. * Use LONG for the reference count of ArchData. * Define PF_XSAVE_ENABLED if it's not already defined by the Windows SDK.

+++ branches/mago64/DebugEngine/MagoNatDE/ArchData.h
@@ -70,7 +70,7 @@ namespace Mago
 
     class ArchData
     {
-        int mRefCount;
+        LONG    mRefCount;
 
     public:
         ArchData();
